requirements.txt
annotated-types==0.7.0
audioop-lts==0.2.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comtypes==1.4.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
gTTS==2.5.4
httplib2==0.31.0
idna==3.10
paho-mqtt==2.1.0
playsound==1.2.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
pypiwin32==223
pyttsx3==2.99
pywin32==311
requests==2.32.5
rsa==4.9.1
sounddevice==0.5.2
SpeechRecognition==3.14.3
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
